removed regions, splitted up classes per file
