+++ CefSharp.Example/CefSharp.Example.csproj
@@ -79,6 +79,22 @@
     <Compile Include="Filters\PassThruResponseFilter.cs" />
     <Compile Include="Handlers\BrowserProcessHandler.cs" />
     <Compile Include="InMemorySchemeAndResourceHandlerFactory.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\BaseRequestEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\GetAuthCredentialsEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\GetResourceResponseFilterEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnBeforeBrowseEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnBeforeResourceLoadEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnCertificateErrorEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnOpenUrlFromTabEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnPluginCrashedEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnProtocolExecutionEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnQuotaRequestEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnRenderProcessTerminatedEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnRenderViewReadyEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnResourceLoadCompleteEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnResourceRedirectEventArgs.cs" />
+    <Compile Include="RequestEventHandler\EventArgs\OnResourceResponseEventArgs.cs" />
+    <Compile Include="RequestEventHandler\RequestEventHandler.cs" />
     <Compile Include="ScriptedMethodsBoundObject.cs" />
     <Compile Include="ExceptionTestBoundObject.cs" />
     <Compile Include="JsDialogHandler.cs" />
