+++ CefSharp.Example/RequestEventHandler/EventArgs/OnCertificateErrorEventArgs.cs
@@ -1,0 +1,27 @@
+﻿namespace CefSharp.Example.RequestEventHandler {
+    public class OnCertificateErrorEventArgs : BaseRequestEventArgs {
+        public OnCertificateErrorEventArgs(IWebBrowser browserControl, IBrowser browser, CefErrorCode errorCode, string requestUrl, ISslInfo sslInfo, IRequestCallback callback)
+            : base(browserControl, browser) {
+            ErrorCode = errorCode;
+            RequestUrl = requestUrl;
+            SSLInfo = sslInfo;
+            Callback = callback;
+        }
+
+        public CefErrorCode ErrorCode { get; }
+        public string RequestUrl { get; }
+        public ISslInfo SSLInfo { get; }
+
+        /// <summary>
+        ///     Callback interface used for asynchronous continuation of url requests.
+        ///     If empty the error cannot be recovered from and the request will be canceled automatically.
+        /// </summary>
+        public IRequestCallback Callback { get; }
+
+        /// <summary>
+        ///     Set to false to cancel the request immediately. Set to true and use <see cref="T:CefSharp.IRequestCallback" /> to
+        ///     execute in an async fashion.
+        /// </summary>
+        public bool ContinueAsync { get; set; } = false;
+    }
+}
