+++ CefSharp.Example/RequestEventHandler/EventArgs/OnPluginCrashedEventArgs.cs
@@ -1,0 +1,9 @@
+﻿namespace CefSharp.Example.RequestEventHandler {
+    public class OnPluginCrashedEventArgs : BaseRequestEventArgs {
+        public OnPluginCrashedEventArgs(IWebBrowser browserControl, IBrowser browser, string pluginPath) : base(browserControl, browser) {
+            PluginPath = pluginPath;
+        }
+
+        public string PluginPath { get; }
+    }
+}
