+++ CefSharp.Example/RequestEventHandler/EventArgs/OnRenderProcessTerminatedEventArgs.cs
@@ -1,0 +1,10 @@
+﻿namespace CefSharp.Example.RequestEventHandler {
+    public class OnRenderProcessTerminatedEventArgs : BaseRequestEventArgs {
+        public OnRenderProcessTerminatedEventArgs(IWebBrowser browserControl, IBrowser browser, CefTerminationStatus status)
+            : base(browserControl, browser) {
+            Status = status;
+        }
+
+        public CefTerminationStatus Status { get; }
+    }
+}
