+++ CefSharp.Example/RequestEventHandler/EventArgs/OnRenderViewReadyEventArgs.cs
@@ -1,0 +1,5 @@
+﻿namespace CefSharp.Example.RequestEventHandler {
+    public class OnRenderViewReadyEventArgs : BaseRequestEventArgs {
+        public OnRenderViewReadyEventArgs(IWebBrowser browserControl, IBrowser browser) : base(browserControl, browser) {}
+    }
+}
